update sqrt syntax in Particles header file and decreased diagnostic precision in pointPlaneDeposit 3D example

+++ include/Particles.h
@@ -229,16 +229,16 @@ public:
     yprevious{nParticles,0.0}, 
     zprevious{nParticles,0.0},
     v{nParticles, 0.0}, 
-    vx{nParticles,std::sqrt(static_cast<float>(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
-		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27))*
+    vx{nParticles,std::sqrtf(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
+		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27)*
     std::cos(getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.theta"))*
     std::sin(getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.phi"))}, 
-    vy{nParticles,std::sqrt(static_cast<float>(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
-		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27))*
+    vy{nParticles,std::sqrtf(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
+		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27)*
     std::sin(getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.theta"))*
     std::sin(getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.phi"))}, 
-    vz{nParticles,std::sqrt(static_cast<float>(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
-		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27))*
+    vz{nParticles,std::sqrtf(2.0*getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.energy_eV")*
+		    1.602e-19/getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")/1.66e-27)*
     std::cos(getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.phi"))}, 
     Z{nParticles},
     amu{nParticles,getVariable_cfg<float> (cfg,"impurityParticleSource.initialConditions.impurity_amu")},
